+++ examples/marketplace-plugin/commands/hello.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-# dummy commands entry - not installed by mind (no mind equivalent)
-echo "hello"